backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pyjwt[crypto]==2.8.0
langchain==0.0.350
langchain-community==0.0.10
langchain-core==0.1.10
transformers==4.35.2
torch==2.1.1
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.16.1
aiohttp==3.9.1
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
jinja2==3.1.2
slowapi==0.1.9
